pyproject.toml
[tool.poetry]
name = "python-flask-tutorials"
version = "0.1.0"
description = ""
authors = ["sudopwr <sudopwr@gmail.com>"]
readme = "README.md"
packages = [{include = "main"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.2"
flask-sqlalchemy = "^3.0.3"
pymysql = "^1.0.3"
cryptography = "^41.0.1"
python-dotenv = "^1.0.0"
pyjwt = "^2.7.0"
google-auth = "^2.21.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
